+++ Dominio.Venda/ProdutoVendido.cs
@@ -4,7 +4,7 @@ namespace Dominio.Venda
     {
         public string Descricao { get; set; }
         public int QuantidadeComprada { get; set; }
-        public int ValorUnitario { get; set; }
+        public decimal ValorUnitario { get; set; }
         public decimal QuantidadePromocional { get; set; }
         public decimal ValorUnitarioPromocional { get; set; }
     }
